refactor(products): interfaces de productos refactorizadas y centralizadas
